test(tx-pool): move service controller tests out of service module
